fix: progress bar refactor out into top level component
