+++ src/Features/Quotas/TeamQuotasOverview/TeamQuotasOverview.module.scss
@@ -36,20 +36,6 @@
   line-height: 0.9375rem;
 }
 
-.coverageBar {
-  height: 1rem;
-  width: 17.625rem;
-  border-radius: 0.5625rem;
-  background-color: var(--cds-ui-02);
-  margin-bottom: 0.25rem;
-}
-
-.coverageFiller {
-  border-radius: 0.5625rem;
-  height: 100%;
-  background-color: var(--cds-support-02);
-}
-
 .detailedHeading {
   font-size: 1.25rem;
   letter-spacing: 0;
