+++ src/Features/Workflows/WorkflowQuotaModalContent/WorkflowQuotaModalContent.module.scss
@@ -46,20 +46,6 @@
   margin-bottom: 1.5rem;
 }
 
-.coverageBar {
-  height: 1.125rem;
-  width: 28.125rem;
-  border-radius: 0.5625rem;
-  background-color: var(--cds-ui-02);
-  margin-bottom: 0.25rem;
-}
-
-.coverageFiller {
-  background-color: var(--cds-support-02);
-  border-radius: 0.5625rem;
-  height: 100%;
-}
-
 .detailedText {
   font-size: 0.75rem;
   letter-spacing: 0.02rem;
